Move specific worker interfaces into the general package. Move exceptions into the exception package.
