+++ airbyte-workers/src/main/java/io/airbyte/workers/RecordSchemaValidator.java
@@ -9,6 +9,8 @@ import com.fasterxml.jackson.databind.node.ObjectNode;
 import io.airbyte.protocol.models.AirbyteRecordMessage;
 import io.airbyte.validation.json.JsonSchemaValidator;
 import io.airbyte.validation.json.JsonValidationException;
+import io.airbyte.workers.exception.RecordSchemaValidationException;
+
 import java.util.Map;
 import java.util.Objects;
