+++ airbyte-workers/src/main/java/io/airbyte/workers/general/DefaultNormalizationWorker.java
@@ -7,8 +7,8 @@ package io.airbyte.workers.general;
 import io.airbyte.config.Configs.WorkerEnvironment;
 import io.airbyte.config.NormalizationInput;
 import io.airbyte.config.NormalizationSummary;
-import io.airbyte.workers.NormalizationWorker;
-import io.airbyte.workers.WorkerException;
+import io.airbyte.workers.normalization.NormalizationWorker;
+import io.airbyte.workers.exception.WorkerException;
 import io.airbyte.workers.normalization.NormalizationRunner;
 import java.nio.file.Files;
 import java.nio.file.Path;
