+++ airbyte-workers/src/main/java/io/airbyte/workers/general/ReplicationWorker.java
@@ -2,9 +2,10 @@
  * Copyright (c) 2021 Airbyte, Inc., all rights reserved.
  */
 
-package io.airbyte.workers;
+package io.airbyte.workers.general;
 
 import io.airbyte.config.ReplicationOutput;
 import io.airbyte.config.StandardSyncInput;
+import io.airbyte.workers.Worker;
 
 public interface ReplicationWorker extends Worker<StandardSyncInput, ReplicationOutput> {}
