+++ airbyte-workers/src/main/java/io/airbyte/workers/helper/EntrypointEnvChecker.java
@@ -4,7 +4,7 @@
 
 package io.airbyte.workers.helper;
 
-import io.airbyte.workers.WorkerException;
+import io.airbyte.workers.exception.WorkerException;
 import io.airbyte.workers.process.ProcessFactory;
 import java.io.BufferedReader;
 import java.io.IOException;
