+++ airbyte-workers/src/main/java/io/airbyte/workers/process/ProcessFactory.java
@@ -5,7 +5,7 @@
 package io.airbyte.workers.process;
 
 import io.airbyte.config.ResourceRequirements;
-import io.airbyte.workers.WorkerException;
+import io.airbyte.workers.exception.WorkerException;
 import java.nio.file.Path;
 import java.util.Map;
 import java.util.regex.Matcher;
