+++ airbyte-workers/src/main/java/io/airbyte/workers/storage/StateClients.java
@@ -5,7 +5,7 @@
 package io.airbyte.workers.storage;
 
 import io.airbyte.config.storage.CloudStorageConfigs;
-import io.airbyte.workers.DocumentStoreClient;
+import io.airbyte.workers.general.DocumentStoreClient;
 import java.nio.file.Path;
 
 public class StateClients {
